+++ c-api/libvmexeccapi.h
@@ -129,6 +129,7 @@ typedef struct {
   void (*managed_get_prev_block_random_seed_func_ptr)(void *context, int32_t result_handle);
   void (*managed_get_return_data_func_ptr)(void *context, int32_t result_id, int32_t result_handle);
   void (*managed_get_multi_dcdt_call_value_func_ptr)(void *context, int32_t multi_call_value_handle);
+  void (*managed_get_back_transfers_func_ptr)(void *context, int32_t dcdt_transfers_value_handle, int32_t call_value_handle);
   void (*managed_get_dcdt_balance_func_ptr)(void *context, int32_t address_handle, int32_t token_id_handle, int64_t nonce, int32_t value_handle);
   void (*managed_get_dcdt_token_data_func_ptr)(void *context, int32_t address_handle, int32_t token_id_handle, int64_t nonce, int32_t value_handle, int32_t properties_handle, int32_t hash_handle, int32_t name_handle, int32_t attributes_handle, int32_t creator_handle, int32_t royalties_handle, int32_t uris_handle);
   void (*managed_async_call_func_ptr)(void *context, int32_t dest_handle, int32_t value_handle, int32_t function_handle, int32_t arguments_handle);
@@ -316,7 +317,7 @@ vm_exec_result_t vm_check_signatures(vm_exec_instance_t *instance_ptr);
  *
  * C API function, works with raw object pointers.
  */
-void vm_exec_executor_destroy(vm_exec_executor_t *executor);
+void vm_exec_executor_destroy(vm_exec_executor_t *executor_ptr);
 
 /**
  * Sets the data that can be hold by an instance context.
@@ -379,7 +380,7 @@ vm_exec_result_t vm_exec_instance_call(vm_exec_instance_t *instance_ptr, const c
  *
  * C API function, works with raw object pointers.
  */
-void vm_exec_instance_destroy(vm_exec_instance_t *instance);
+void vm_exec_instance_destroy(vm_exec_instance_t *instance_ptr);
 
 /**
  * Creates a new VM executor instance from cache.