+++ c-api/src/capi_vm_hooks.rs
@@ -403,6 +403,10 @@ impl dharitri_vm_executor::VMHooks for CapiVMHooks {
         (self.c_func_pointers_ptr.managed_get_multi_dcdt_call_value_func_ptr)(self.vm_hooks_ptr, multi_call_value_handle)
     }
 
+    fn managed_get_back_transfers(&self, dcdt_transfers_value_handle: i32, call_value_handle: i32) {
+        (self.c_func_pointers_ptr.managed_get_back_transfers_func_ptr)(self.vm_hooks_ptr, dcdt_transfers_value_handle, call_value_handle)
+    }
+
     fn managed_get_dcdt_balance(&self, address_handle: i32, token_id_handle: i32, nonce: i64, value_handle: i32) {
         (self.c_func_pointers_ptr.managed_get_dcdt_balance_func_ptr)(self.vm_hooks_ptr, address_handle, token_id_handle, nonce, value_handle)
     }